feat: 미네랄 들고 버리기 구현

+++ SpaceMuseum/Assets/Script/Mineral/MineralData.cs
@@ -8,7 +8,7 @@ public class MineralData : ScriptableObject
     public Material mineralMaterial; // 자원 큐브에 적용할 색상 Material
 
     [Header("게임 데이터")]
-    public float weight;            // 자원의 무게
+    public int weight;            // 자원의 무게
     public int price;               // 자원의 판매 가격
     public float spawnProbability;  // 상자에서 등장할 확률 (가중치)
 }
